requirements/base.txt
psycopg2-binary
celery>=4.1.0
django-celery-beat
python-memcached
django-sekizai
django_compressor
csscompressor
beautifulsoup4
lxml
django-htmlmin>=0.11.0
-e git+git@github.com:nargeel/django-fancy-imagefield.git@master#egg=django-fancy-imagefield
-e git+git@github.com:nargeel/django-contrib.git@master#egg=django-contrib
django-cleanup
django-redis-sessionsdjango-constance==2.4.0
